Moderation UI + minor backend update
+Added stylized action buttons to Moderation tab
+Added columns to moderation log table

+Added ModerationRecords datatype
+Added getModRecord and getModRecordCount to requests.js

I haven't been able to populate the table with data yet. Some vestiges from my attempts are still in this commit, so don't merge until these (mostly in OrganizationModeration.jsx) are handled somehow.
I will prioritize fixing this in the next commit.

A readme to explain some of the changes will also be added in the next commit.